+++ bstore/parsers.py
@@ -11,6 +11,7 @@ import sys
 import tkinter as tk
 import bstore.database as db
 import traceback
+import warnings
 
 __version__ = config.__bstore_Version__
 
@@ -175,10 +176,6 @@ class PositionParser(Parser):
             # Save the full path to the file for later.
             # If filename is already a Path object, this does nothing.
             self._fullPath = pathlib.Path(filename)        
-            
-            # Convert Path objects to strings if Path is supplied
-            #if isinstance(filename, pathlib.PurePath):
-            #    filename = str(filename.name)
     
             # Remove file type ending and any parent folders
             # Example: 'path/to/HeLa_Control_7.csv' becomes 'HeLa_Control_7'
@@ -192,11 +189,18 @@ class PositionParser(Parser):
                                                                   datasetType))
             dType             = getattr(mod, datasetType)
             self.dataset      = dType(datasetIDs = idDict)
-            self.dataset.data = self.dataset.readFromFile(self._fullPath)
         except:
             if config.__Verbose__:
                 print(traceback.format_exc())
             raise ParseFilenameFailure('ParseFilenameError')
+        
+        # Read the data from file
+        try:
+            self.dataset.data = self.dataset.readFromFile(
+                                                 self._fullPath, **kwargs)
+        except:
+            warnings.warn('Warning: Filename successfully parsed, but no data '
+                          'was read from the file.')
                                         
     def _parse(self, rootName):
         """Actually does the work of splitting the name and finding IDs.
@@ -211,7 +215,14 @@ class PositionParser(Parser):
         
         """
         idDict = {}
-        for pos, field in enumerate(rootName.split(self.sep)):
+        fields = rootName.split(self.sep)
+        
+        if max(self.positionIDs.keys()) + 1 > len(fields):
+            # Raises an error if more fields are supplied than are present in
+            # the filename.
+            raise ParseFilenameFailure('ParseFilenameError')
+        
+        for pos, field in enumerate(fields):
             # Skip empty or positions or those marked None
             if pos not in self.positionIDs or self.positionIDs[pos] is None:
                 continue
@@ -423,11 +434,18 @@ class SimpleParser(Parser):
                 'bstore.datasetTypes.{0:s}'.format(datasetType))
             dType             = getattr(mod, datasetType)
             self.dataset      = dType(datasetIDs = idDict)
-            self.dataset.data = self.dataset.readFromFile(self._fullPath)
         except:
             self.dataset = None
             raise ParseFilenameFailure(('Error: File could not be parsed.',
                                         sys.exc_info()[0]))
+                                        
+        # Read the data from file
+        try:
+            self.dataset.data = self.dataset.readFromFile(
+                                                 self._fullPath, **kwargs)
+        except:
+            warnings.warn('Warning: Filename successfully parsed, but no data '
+                          'was read from the file.')
 
 """
 Exceptions